+++ goat/command/build_command/factory/build_command_factory.py
@@ -1,7 +1,7 @@
-from goat.command.build_command.builder.base_build_command_factory import (
+from goat.command.build_command.factory.base_build_command_factory import (
     BaseBuildCommandFactory,
 )
-from goat.command.build_command.builder.gpp_build_command_factory import (
+from goat.command.build_command.factory.gpp_build_command_factory import (
     GPPBuildCommandFactory,
 )
 from goat.command.build_command.parameters.compile_parameters import CompileParameters
